Make embedded document serialization canonical JSON.
It is expected by the Rails app that a document may be entirely
initialized from a JSON-structured hash.  Therefore, we finally bit the
bullet and implemented recursive parsing built into attribute assignment
that will automatically covert a hash of attributes to its corresponding
embedded document type.  With that in place, serialization was then
switch over to the canonical JSON structure.

This change also applied to `many' embedded document sequences.