Attempt to clarify ordered_by's :count.
Prior to this commit, :count was inaccurately described as an upper
limit on the number of documents returned.  While this may be the case
in some cases, it is not true for all of them.  The ordering is
maintained as a supercolumn family ordered on a given document field.
Document keys are stored in groups that share the same value for the
ordered field.  The :count option establishes an upper limit on the
number of *groups* that are returned.  This is only the same as the
number of documents when each document has a unique value for the
ordering field.