Add pre-obliteration callbacks.
A versioned document is essentially not truly destroyed until it becomes
an expired zombie.  The `before_destroy' callbacks only trigger when the
document is initially turned into a zombie.  However, some circumstances
require additional processing when a document is finally, truly deleted.

Therefore, a `before_obliterate' set of callbacks has been provided that
is triggered just before a document is truly destroyed (either by
becoming an expired zombie or by a call to `permanently_destroy').