Increment version numbers before saving.
If version numbers are incremented after saving, then incorrect versions
will be used when reloading and resaving a document from the database;
the version number would need to increment but wouldn't until after the
save.

Therefore, version numbers are incremented prior to saving, and a new,
not-yet-saved object starts with a version number of 0 (which will
become 1 just before it is written to the database).