Fix: Remove unintended modifications in serialization.
The calls to `force_encoding' actually modify the string they are
invoked on instead of returning a new version of that string with
modified encoding.  Because the string objects passed into serialization
are often part of a larger data structure (usually, the _raw_columns hash),
these changes propagate all the way up to that original data structure.

This was resolved by `dup'ing the strings before changing their
encoding.